+++ myems-admin/app/controllers/settings/virtualpowerplant/virtualpowerplant.controller.js
@@ -12,6 +12,9 @@ app.controller('VirtualPowerPlantController', function(
     toaster,
     SweetAlert) {
 	$scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
+	$scope.exportdata = '';
+	$scope.importdata = '';
+
 	$scope.getAllVirtualPowerPlants = function() {
 		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
 		VirtualPowerPlantService.getAllVirtualPowerPlants(headers, function (response) {
@@ -166,6 +169,97 @@ app.controller('VirtualPowerPlantController', function(
 		        }
 		    }
 		);
+	};
+
+	$scope.exportVirtualPowerPlant = function(virtualpowerplant) {
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		VirtualPowerPlantService.exportVirtualPowerPlant(virtualpowerplant, headers, function(response) {
+			if (angular.isDefined(response.status) && response.status === 200) {
+				$scope.exportdata = JSON.stringify(response.data);
+				var modalInstance = $uibModal.open({
+					windowClass: "animated fadeIn",
+					templateUrl: 'views/common/export.html',
+					controller: 'ModalExportCtrl',
+					resolve: {
+						params: function() {
+							return {
+								exportdata: angular.copy($scope.exportdata)
+							};
+						}
+					}
+				});
+				modalInstance.result.then(function() {
+					//do nothing;
+				}, function() {
+					//do nothing;
+				});
+				$rootScope.modalInstance = modalInstance;
+			} else {
+				$scope.exportdata = null;
+			}
+		});
+	};
+
+	$scope.cloneVirtualPowerPlant = function(virtualpowerplant){
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		VirtualPowerPlantService.cloneVirtualPowerPlant(virtualpowerplant, headers, function(response) {
+			if (angular.isDefined(response.status) && response.status === 201) {
+				toaster.pop({
+					type: "success",
+					title: $translate.instant("TOASTER.SUCCESS_TITLE"),
+					body: $translate.instant("TOASTER.SUCCESS_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+					showCloseButton: true,
+				});
+				$scope.getAllVirtualPowerPlants();
+				$scope.$emit('handleEmitVirtualPowerPlantChanged');
+			}else {
+				toaster.pop({
+					type: "error",
+					title: $translate.instant("TOASTER.ERROR_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+					body: $translate.instant(response.data.description),
+					showCloseButton: true,
+				});
+			}
+		});
+	};
+
+	$scope.importVirtualPowerPlant = function() {
+		var modalInstance = $uibModal.open({
+			templateUrl: 'views/common/import.html',
+			controller: 'ModalImportCtrl',
+			windowClass: "animated fadeIn",
+			resolve: {
+				params: function() {
+					return {
+					};
+				}
+			}
+		});
+		modalInstance.result.then(function(importdata) {
+			let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+			VirtualPowerPlantService.importVirtualPowerPlant(importdata, headers, function(response) {
+				if (angular.isDefined(response.status) && response.status === 201) {
+					toaster.pop({
+						type: "success",
+						title: $translate.instant("TOASTER.SUCCESS_TITLE"),
+						body: $translate.instant("TOASTER.SUCCESS_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+						showCloseButton: true,
+					});
+					$scope.getAllVirtualPowerPlants();
+					$scope.$emit('handleEmitVirtualPowerPlantChanged');
+				} else {
+					toaster.pop({
+						type: "error",
+						title: $translate.instant("TOASTER.ERROR_ADD_BODY", { template: $translate.instant("COMMON.DATA_SOURCE") }),
+						body: $translate.instant(response.data.description),
+						showCloseButton: true,
+					});
+				}
+			});
+		}, function() {
+
+		});
+		$rootScope.modalInstance = modalInstance;
 	};
 
 	$scope.getAllVirtualPowerPlants();