+++ myems-api/reports/virtualmetercarbon.py
@@ -3,7 +3,6 @@ from decimal import Decimal
 import falcon
 import mysql.connector
 import simplejson as json
-
 import config
 import excelexporters.virtualmetercarbon
 from core import utilities